Fix clean mails before scenarios hook

+++ features/support/env.rb
@@ -36,8 +36,10 @@ end
 Before do
   # clean mails after running tests
   mail_store = "#{Padrino.root}/tmp/emails"
-  Dir.each_child(mail_store) do |file|
-    File.delete("#{mail_store}/#{file}")
+  if Dir.exist?(mail_store)
+    Dir.each_child(mail_store) do |file|
+      File.delete("#{mail_store}/#{file}")
+    end
   end
 end
